+++ example/lib/main.dart
@@ -55,7 +55,7 @@ class _MyAppState extends State<MyApp> with SingleTickerProviderStateMixin {
               ),
 
               // The end action pane is the one at the right or the bottom side.
-              endActionPane:  ActionPane(
+              endActionPane: ActionPane(
                 motion: const ScrollMotion(),
                 children: [
                   SlidableAction(