+++ lib/src/action_pane_motions.dart
@@ -11,8 +11,8 @@ class BehindMotion extends StatelessWidget {
   ///
   /// {@animation 664 200 https://raw.githubusercontent.com/letsar/flutter_slidable/assets/behind_motion.mp4}
   const BehindMotion({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -31,8 +31,8 @@ class StretchMotion extends StatelessWidget {
   ///
   /// {@animation 664 200 https://raw.githubusercontent.com/letsar/flutter_slidable/assets/stretch_motion.mp4}
   const StretchMotion({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -63,8 +63,8 @@ class ScrollMotion extends StatelessWidget {
   ///
   /// {@animation 664 200 https://raw.githubusercontent.com/letsar/flutter_slidable/assets/scroll_motion.mp4}
   const ScrollMotion({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -91,8 +91,8 @@ class DrawerMotion extends StatelessWidget {
   ///
   /// {@animation 664 200 https://raw.githubusercontent.com/letsar/flutter_slidable/assets/drawer_motion.mp4}
   const DrawerMotion({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {