+++ lib/src/auto_close_behavior.dart
@@ -8,11 +8,11 @@ import 'package:flutter_slidable/src/notifications.dart';
 class SlidableAutoCloseBehavior extends StatefulWidget {
   /// Creates a [SlidableAutoCloseBehavior].
   const SlidableAutoCloseBehavior({
-    Key? key,
+    super.key,
     this.closeWhenOpened = true,
     this.closeWhenTapped = true,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// Indicates whether all the [Slidable] within the same group should be
   /// closed when one of the group is opened.
@@ -67,11 +67,10 @@ class _SlidableAutoCloseBehaviorState extends State<SlidableAutoCloseBehavior> {
 
 class _SlidableAutoCloseData extends InheritedWidget {
   const _SlidableAutoCloseData({
-    Key? key,
     required this.closeWhenOpened,
     required this.closeWhenTapped,
-    required Widget child,
-  }) : super(key: key, child: child);
+    required super.child,
+  });
 
   final bool closeWhenOpened;
   final bool closeWhenTapped;
@@ -91,11 +90,11 @@ class _SlidableAutoCloseData extends InheritedWidget {
 class SlidableAutoCloseBehaviorInteractor extends StatelessWidget {
   /// INTERNAL USE
   const SlidableAutoCloseBehaviorInteractor({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -150,11 +149,11 @@ class SlidableAutoCloseNotification {
 class SlidableAutoCloseInteractor extends StatelessWidget {
   /// INTERNAL USE
   const SlidableAutoCloseInteractor({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -185,11 +184,11 @@ class SlidableAutoCloseInteractor extends StatelessWidget {
 class SlidableAutoCloseBehaviorListener extends StatelessWidget {
   /// INTERNAL USE
   const SlidableAutoCloseBehaviorListener({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -221,11 +220,11 @@ class SlidableAutoCloseBehaviorListener extends StatelessWidget {
 class SlidableAutoCloseNotificationSender extends StatelessWidget {
   /// INTERNAL USE
   const SlidableAutoCloseNotificationSender({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -291,11 +290,11 @@ class SlidableAutoCloseBarrierNotification {
 class SlidableAutoCloseBarrierInteractor extends StatelessWidget {
   /// INTERNAL USE
   const SlidableAutoCloseBarrierInteractor({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -326,11 +325,11 @@ class SlidableAutoCloseBarrierInteractor extends StatelessWidget {
 class SlidableAutoCloseBarrierNotificationSender extends StatefulWidget {
   /// INTERNAL USE
   const SlidableAutoCloseBarrierNotificationSender({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -417,11 +416,11 @@ class _SlidableAutoCloseBarrierNotificationSenderState
 class SlidableAutoCloseBarrierBehaviorListener extends StatefulWidget {
   /// INTERNAL USE
   const SlidableAutoCloseBarrierBehaviorListener({
-    Key? key,
+    super.key,
     required this.groupTag,
     required this.controller,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// {@macro slidable.groupTag}
   final Object? groupTag;
@@ -484,12 +483,11 @@ class _SlidableAutoCloseBarrierBehaviorListenerState
 class _SlidableNotificationSender extends StatefulWidget {
   /// INTERNAL USE
   const _SlidableNotificationSender({
-    Key? key,
     required this.controller,
     required this.onStatusChanged,
     required this.enabled,
     required this.child,
-  }) : super(key: key);
+  });
 
   final SlidableController controller;